requirements.txt
gradio
tensorflow
pillow
numpy
gdown